requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
celery[redis]==5.3.1
aerich==0.9.1
tortoise-orm==0.25.1
asyncpg==0.30.0
pytest==7.4.0
pytest-asyncio==0.21.1
boto3==1.35.0
replicate==0.25.1
httpx==0.27.0
Pillow==10.0.0
loguru==0.7.2